backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pymupdf==1.23.21
langchain==0.1.6
tiktoken==0.5.2
openai==1.12.0
sentence-transformers==2.3.1
python-dotenv==1.0.1
requests==2.31.0
numpy==1.26.4
